extruct worker_sender
